// src/models/IPhones.ts
interface IPhones {
  id: number;
  name: string;
  company: string;
  short_desc: string;
  img: string;
  threads_name: string;
  front_camera: string;
  main_camera: string;
  count: number;
  quantity_threads: number;
  memory: number;
  price: number;
  click: boolean;
  click_versus: boolean;
}